♻️ [2025-07-29] 重构代码: #43 把 client base，核心方法放入 core 文件夹中
