+++ custom_components/lifesmart/core/__init__.py
@@ -1,0 +1,1 @@
+# 'core'目录是一个核心包
