+++ custom_components/lifesmart/core/lifesmart_client.py
@@ -15,39 +15,22 @@ from aiohttp.client_exceptions import ClientError
 from homeassistant.core import HomeAssistant
 from homeassistant.helpers.aiohttp_client import async_get_clientsession
 
-from .const import (
+from custom_components.lifesmart.const import (
     # --- 命令类型常量 ---
-    CMD_TYPE_ON,
-    CMD_TYPE_OFF,
-    CMD_TYPE_PRESS,
-    CMD_TYPE_SET_VAL,
-    CMD_TYPE_SET_CONFIG,
-    CMD_TYPE_SET_TEMP_DECIMAL,
-    CMD_TYPE_SET_RAW,
-    CMD_TYPE_SET_TEMP_FCU,
+    # (这些常量现在由基类使用)
     # --- 设备类型和映射 ---
     HUB_ID_KEY,
     DEVICE_ID_KEY,
     SUBDEVICE_INDEX_KEY,
-    DOOYA_TYPES,
-    GARAGE_DOOR_TYPES,
-    NON_POSITIONAL_COVER_CONFIG,
-    REVERSE_F_HVAC_MODE_MAP,
-    LIFESMART_F_FAN_MAP,
-    LIFESMART_TF_FAN_MAP,
-    LIFESMART_ACIPM_FAN_MAP,
-    LIFESMART_CP_AIR_FAN_MAP,
-    REVERSE_LIFESMART_HVAC_MODE_MAP,
-    REVERSE_LIFESMART_CP_AIR_HVAC_MODE_MAP,
-    HVACMode,
 )
-from .diagnostics import get_error_advice
-from .exceptions import LifeSmartAPIError, LifeSmartAuthError
+from custom_components.lifesmart.core.client_base import LifeSmartClientBase
+from custom_components.lifesmart.diagnostics import get_error_advice
+from custom_components.lifesmart.exceptions import LifeSmartAPIError, LifeSmartAuthError
 
 _LOGGER = logging.getLogger(__name__)
 
 
-class LifeSmartClient:
+class LifeSmartClient(LifeSmartClientBase):
     """一个用于高效、健壮地管理 LifeSmart API 调用的类。
 
     此类封装了所有与 LifeSmart API 的通信细节，包括认证、签名生成、
@@ -421,196 +404,32 @@ class LifeSmartClient:
         return self._get_code_from_response(response, "SendKeys")
 
     # ====================================================================
-    # 设备直接控制的辅助方法
+    # 基类抽象方法的实现
     # ====================================================================
 
-    async def turn_on_light_switch_async(self, idx: str, agt: str, me: str) -> int:
-        """开启一个灯或开关。"""
-        return await self.set_single_ep_async(agt, me, idx, CMD_TYPE_ON, 1)
-
-    async def turn_off_light_switch_async(self, idx: str, agt: str, me: str) -> int:
-        """关闭一个灯或开关。"""
-        return await self.set_single_ep_async(agt, me, idx, CMD_TYPE_OFF, 0)
-
-    async def press_switch_async(
-        self, idx: str, agt: str, me: str, duration_ms: int
+    async def _async_send_single_command(
+        self, agt: str, me: str, idx: str, command_type: str, val: Any
     ) -> int:
-        """执行点动操作（按下后在指定时间后自动弹起）。"""
-        # 将毫秒转换为设备需要的 100ms 单位
-        val = max(1, round(duration_ms / 100))
-        return await self.set_single_ep_async(agt, me, idx, CMD_TYPE_PRESS, val)
+        """
+        [云端实现] 发送单个IO口命令。
+        此方法通过调用 set_single_ep_async 来实现基类的抽象方法。
+        """
+        return await self.set_single_ep_async(agt, me, idx, command_type, val)
 
-    async def open_cover_async(self, agt: str, me: str, device_type: str) -> int:
-        """开启窗帘或车库门。"""
-        # 优先处理有特殊命令的定位窗帘
-        if device_type in GARAGE_DOOR_TYPES:
-            return await self.set_single_ep_async(agt, me, "P3", CMD_TYPE_SET_VAL, 100)
-        if device_type in DOOYA_TYPES:
-            return await self.set_single_ep_async(agt, me, "P2", CMD_TYPE_SET_VAL, 100)
-
-        # 对于非定位窗帘，从 const 配置中查找 'open' 对应的 IO 口
-        if device_type in NON_POSITIONAL_COVER_CONFIG:
-            cmd_idx = NON_POSITIONAL_COVER_CONFIG[device_type]["open"]
-            return await self.set_single_ep_async(agt, me, cmd_idx, CMD_TYPE_ON, 1)
-
-        _LOGGER.warning("设备类型 %s 的 'open_cover' 操作未被支持。", device_type)
-        return -1
-
-    async def close_cover_async(self, agt: str, me: str, device_type: str) -> int:
-        """关闭窗帘或车库门。"""
-        # 优先处理有特殊命令的定位窗帘
-        if device_type in GARAGE_DOOR_TYPES:
-            return await self.set_single_ep_async(agt, me, "P3", CMD_TYPE_SET_VAL, 0)
-        if device_type in DOOYA_TYPES:
-            return await self.set_single_ep_async(agt, me, "P2", CMD_TYPE_SET_VAL, 0)
-
-        # 对于非定位窗帘，从 const 配置中查找 'close' 对应的 IO 口
-        if device_type in NON_POSITIONAL_COVER_CONFIG:
-            cmd_idx = NON_POSITIONAL_COVER_CONFIG[device_type]["close"]
-            return await self.set_single_ep_async(agt, me, cmd_idx, CMD_TYPE_ON, 1)
-
-        _LOGGER.warning("设备类型 %s 的 'close_cover' 操作未被支持。", device_type)
-        return -1
-
-    async def stop_cover_async(self, agt: str, me: str, device_type: str) -> int:
-        """停止窗帘或车库门。"""
-        # 定位窗帘的停止命令是特殊的
-        if device_type in GARAGE_DOOR_TYPES:
-            return await self.set_single_ep_async(
-                agt, me, "P3", CMD_TYPE_SET_CONFIG, CMD_TYPE_OFF
-            )
-        if device_type in DOOYA_TYPES:
-            return await self.set_single_ep_async(
-                agt, me, "P2", CMD_TYPE_SET_CONFIG, CMD_TYPE_OFF
-            )
-
-        # 对于非定位窗帘，从 const 配置中查找 'stop' 对应的 IO 口
-        if device_type in NON_POSITIONAL_COVER_CONFIG:
-            cmd_idx = NON_POSITIONAL_COVER_CONFIG[device_type]["stop"]
-            return await self.set_single_ep_async(agt, me, cmd_idx, CMD_TYPE_ON, 1)
-
-        _LOGGER.warning("设备类型 %s 的 'stop_cover' 操作未被支持。", device_type)
-        return -1
-
-    async def set_cover_position_async(
-        self, agt: str, me: str, position: int, device_type: str
+    async def _async_send_multi_command(
+        self, agt: str, me: str, io_list: list[dict]
     ) -> int:
-        """设置窗帘或车库门到指定位置。"""
-        if device_type in GARAGE_DOOR_TYPES:
-            return await self.set_single_ep_async(
-                agt, me, "P3", CMD_TYPE_SET_VAL, position
-            )
-        elif device_type in DOOYA_TYPES:
-            return await self.set_single_ep_async(
-                agt, me, "P2", CMD_TYPE_SET_VAL, position
-            )
-        _LOGGER.warning("设备类型 %s 不支持设置位置。", device_type)
-        return -1
+        """
+        [云端实现] 同时发送多个IO口的命令。
+        此方法通过调用 set_multi_eps_async 来实现基类的抽象方法。
+        """
+        return await self.set_multi_eps_async(agt, me, io_list)
 
-    async def async_set_climate_hvac_mode(
-        self,
-        agt: str,
-        me: str,
-        device_type: str,
-        hvac_mode: HVACMode,
-        current_val: int = 0,
-    ) -> int:
-        """设置温控设备的风扇模式。"""
-        if hvac_mode == HVACMode.OFF:
-            return await self.set_single_ep_async(agt, me, "P1", CMD_TYPE_OFF, 0)
-
-        await self.set_single_ep_async(agt, me, "P1", CMD_TYPE_ON, 1)
-
-        mode_val = None
-        idx = None
-
-        if device_type == "V_AIR_P":
-            mode_val = REVERSE_F_HVAC_MODE_MAP.get(hvac_mode)
-            idx = "MODE"
-        elif device_type in {"SL_NATURE", "SL_FCU"}:
-            mode_val = REVERSE_LIFESMART_HVAC_MODE_MAP.get(hvac_mode)
-            idx = "P7"
-        elif device_type == "SL_UACCB":
-            mode_val = REVERSE_LIFESMART_HVAC_MODE_MAP.get(hvac_mode)
-            idx = "P2"
-
-        if mode_val is not None and idx is not None:
-            return await self.set_single_ep_async(
-                agt, me, idx, CMD_TYPE_SET_CONFIG, mode_val
-            )
-
-        if device_type == "SL_CP_AIR":
-            mode_val = REVERSE_LIFESMART_CP_AIR_HVAC_MODE_MAP.get(hvac_mode)
-            if mode_val is not None:
-                new_val = (current_val & ~(0b11 << 13)) | (mode_val << 13)
-                return await self.set_single_ep_async(
-                    agt, me, "P1", CMD_TYPE_SET_RAW, new_val
-                )
-
-        if device_type == "SL_CP_DN":
-            is_auto = 1 if hvac_mode == HVACMode.AUTO else 0
-            new_val = (current_val & ~(1 << 31)) | (is_auto << 31)
-            return await self.set_single_ep_async(
-                agt, me, "P1", CMD_TYPE_SET_RAW, new_val
-            )
-
-        if device_type == "SL_CP_VL":
-            mode_map = {HVACMode.HEAT: 0, HVACMode.AUTO: 2}
-            mode_val = mode_map.get(hvac_mode, 0)
-            new_val = (current_val & ~(0b11 << 1)) | (mode_val << 1)
-            return await self.set_single_ep_async(
-                agt, me, "P1", CMD_TYPE_SET_RAW, new_val
-            )
-
-        return 0
-
-    async def async_set_climate_temperature(
-        self, agt: str, me: str, device_type: str, temp: float
-    ) -> int:
-        """设置温控设备的目标温度。"""
-        temp_val = int(temp * 10)
-        idx_map = {
-            "V_AIR_P": ("tT", CMD_TYPE_SET_TEMP_DECIMAL),
-            "SL_UACCB": ("P3", CMD_TYPE_SET_TEMP_DECIMAL),
-            "SL_CP_DN": ("P3", CMD_TYPE_SET_RAW),
-            "SL_CP_AIR": ("P4", CMD_TYPE_SET_RAW),
-            "SL_NATURE": ("P8", CMD_TYPE_SET_TEMP_DECIMAL),
-            "SL_FCU": ("P8", CMD_TYPE_SET_TEMP_FCU),
-            "SL_CP_VL": ("P3", CMD_TYPE_SET_RAW),
-        }
-        if device_type in idx_map:
-            idx, cmd_type = idx_map[device_type]
-            return await self.set_single_ep_async(agt, me, idx, cmd_type, temp_val)
-        return -1
-
-    async def async_set_climate_fan_mode(
-        self, agt: str, me: str, device_type: str, fan_mode: str, current_val: int = 0
-    ) -> int:
-        """设置温控设备的风扇模式。"""
-        if device_type == "V_AIR_P":
-            if (fan_val := LIFESMART_F_FAN_MAP.get(fan_mode)) is not None:
-                return await self.set_single_ep_async(
-                    agt, me, "F", CMD_TYPE_SET_CONFIG, fan_val
-                )
-        elif device_type == "SL_TR_ACIPM":
-            if (fan_val := LIFESMART_ACIPM_FAN_MAP.get(fan_mode)) is not None:
-                return await self.set_single_ep_async(
-                    agt, me, "P2", CMD_TYPE_SET_RAW, fan_val
-                )
-        elif device_type in {"SL_NATURE", "SL_FCU"}:
-            if (fan_val := LIFESMART_TF_FAN_MAP.get(fan_mode)) is not None:
-                return await self.set_single_ep_async(
-                    agt, me, "P9", CMD_TYPE_SET_CONFIG, fan_val
-                )
-        elif device_type == "SL_CP_AIR":
-            if (fan_val := LIFESMART_CP_AIR_FAN_MAP.get(fan_mode)) is not None:
-                new_val = (current_val & ~(0b11 << 15)) | (fan_val << 15)
-                return await self.set_single_ep_async(
-                    agt, me, "P1", CMD_TYPE_SET_RAW, new_val
-                )
-
-        _LOGGER.warning("设备类型 %s 不支持风扇模式: %s", device_type, fan_mode)
-        return -1
+    # ====================================================================
+    # 设备直接控制的辅助方法控制方法 (`turn_on_light_switch_async`, `open_cover_async`,
+    # `async_set_climate_hvac_mode` 等) 现已移至 `client_base.py` 中，
+    # 并由该类继承。
+    # ====================================================================
 
     # ====================================================================
     # 内部工具方法
